+++ app/src/main/java/ru/project/iidea/ProjectNotHostViewFragmentInterface.java
@@ -1,6 +1,6 @@
 package ru.project.iidea;
 
 public interface ProjectNotHostViewFragmentInterface {
-    public void onUserIdClicked(long userID);
+    public void onUserIdClicked(int userID);
     public void onBackButtonPressed();
 }
